Expose fake time provider for sync tests
