+++ ShuffleTask.Tests/TestDoubles/StorageServiceStubTests.cs
@@ -11,23 +11,6 @@ namespace ShuffleTask.Tests.TestDoubles;
 [TestFixture]
 public class StorageServiceStubTests
 {
-    private sealed class FakeTimeProvider : TimeProvider
-    {
-        private DateTimeOffset _utcNow;
-
-        public FakeTimeProvider(DateTimeOffset utcNow)
-        {
-            _utcNow = utcNow;
-        }
-
-        public override DateTimeOffset GetUtcNow() => _utcNow;
-
-        public void AdvanceTime(TimeSpan amount)
-        {
-            _utcNow = _utcNow.Add(amount);
-        }
-    }
-
     [Test]
     public async Task InitializeAsync_CanBeCalledMultipleTimes()
     {
